cursorTools/requirements.txt
easyocr>=1.7.1
translate>=3.6.1
Pillow>=10.0.0 